Refactor frontend tests (single everware instance per test)
