+++ build_tools/frontend_test_nonstop_config.py
@@ -1,0 +1,3 @@
+c = get_config()
+load_subconfig('build_tools/frontend_test_normal_config.py')
+c.Spawner.remove_containers = False
