--- build_tools/nonstop_frontend_test_config.py
@@ -1,7 +1,0 @@
-c = get_config()
-load_subconfig('etc/base_config.py')
-
-c.JupyterHub.authenticator_class = 'dummyauthenticator.DummyAuthenticator'
-c.Spawner.start_timeout = 50
-c.Spawner.http_timeout = 120 # docker sometimes doesn't show up for a while
-c.Spawner.remove_containers = False
