+++ frontend_tests/settings.py
@@ -1,0 +1,6 @@
+import os
+
+DRIVER = "phantomjs" # "firefox"
+DEFAULT_REPO = "https://github.com/everware/everware-cpp-example.git"
+TIMEOUT = 250
+UPLOADDIR = os.environ.get('UPLOADDIR')
